VCST-2521: Add release-hotfix workflow (#1698)
## Description
Add a new workflow to create a patch version release. The workflow
increments the patch version in the 'package.json' file in the current
branch only; no merges to other branches are made automatically. The
code changes must be moved to the 'dev' branch manually. The Git Hub
release is created as NOT latest.
## References
### Jira-link:
<!-- Put link to your task in Jira here -->
### Artifact URL:

https://vc3prerelease.blob.core.windows.net/packages/vc-theme-b2b-vue-2.20.0-pr-1698-74af-74af3516.zip